GameHelper to different file
